feat: switch default screen from dashboard to markets
